clean up Main function
